+++ src/constants.ts
@@ -1,3 +1,4 @@
 import type { InitalSelectionOption } from './types';
 
 export const initialSelectionOptionDefault: InitalSelectionOption = 'Nothing';
+export const defaultMaxDepth = 3;