+++ changelog.md
@@ -3,6 +3,7 @@
 ## 1.4.0
 ### Added
 - Setting for whether the first item should be selected when opening the palette or not (#12)
+- Setting configuring the max. depth of the displayed items (#14)
 
 
 ## 1.3.1